relocates files into room_scaper package
